Refactor posts query into helper function
